# perf: add test to run json events