--- a/perf/perf_json.py.data/perf_json.yaml
+++ b/perf/perf_json.py.data/perf_json.yaml
@@ -1,0 +1,6 @@
+run_type: !mux
+    distro:
+        type: 'distro'
+    upstream:
+        location: 'https://github.com/torvalds/linux/archive/master.zip'
+        type: 'upstream'