This patch adds a new test case to run all the perf pmu events. Also compares the event list and event code between perf command and json files from kernel source.

Result on RH8.x
avocado run --max-parallel-tasks=1 perf_json.py
JOB ID     : afb342cdee4ec378bcd801445c82977563930073
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2024-03-15T03.09-afb342c/job.log
 (1/2) perf_json.py:perf_json.test_pmu_events: STARTED
 (1/2) perf_json.py:perf_json.test_pmu_events:  PASS (2755.50 s)
 (2/2) perf_json.py:perf_json.test_compare: STARTED
 (2/2) perf_json.py:perf_json.test_compare:  PASS (2738.79 s)
RESULTS    : PASS 2 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2024-03-15T03.09-afb342c/results.html
JOB TIME   : 5505.32 s

Result on S15 SPx
avocado run --max-parallel-tasks=1 perf_json.py
JOB ID     : 72648f662566c1f901c2d477babe786c4344a07d
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2024-02-29T18.42-72648f6/job.log
 (1/2) perf_json.py:perf_json.test_pmu_events: STARTED
 (1/2) perf_json.py:perf_json.test_pmu_events:  PASS (734.48 s)
 (2/2) perf_json.py:perf_json.test_compare: STARTED
 (2/2) perf_json.py:perf_json.test_compare:  PASS (652.49 s)
RESULTS    : PASS 2 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2024-02-29T18.42-72648f6/results.html
JOB TIME   : 1556.65 s

[json debug2.log](https://github.com/avocado-framework-tests/avocado-misc-tests/files/14617422/json.debug2.log)
[json debug1.log](https://github.com/avocado-framework-tests/avocado-misc-tests/files/14617421/json.debug1.log)