@disgoel 2024-04-24
```
avocado run --max-parallel-tasks=1 perf_json.py -m perf_json.py.data/perf_json.yaml
Fetching asset from perf_json.py:perf_json.test_pmu_events
Fetching asset from perf_json.py:perf_json.test_compare
JOB ID     : 774a6d5c797027cc58debb9600c6f00a0de8e40c
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2024-04-23T14.24-774a6d5/job.log
 (1/4) perf_json.py:perf_json.test_pmu_events;run-run_type-distro-8228: STARTED
 (1/4) perf_json.py:perf_json.test_pmu_events;run-run_type-distro-8228:  PASS (2778.61 s)
 (2/4) perf_json.py:perf_json.test_pmu_events;run-run_type-upstream-3987: STARTED
 (2/4) perf_json.py:perf_json.test_pmu_events;run-run_type-upstream-3987:  PASS (694.42 s)
 (3/4) perf_json.py:perf_json.test_compare;run-run_type-distro-8228: STARTED
 (3/4) perf_json.py:perf_json.test_compare;run-run_type-distro-8228:  PASS (2779.07 s)
 (4/4) perf_json.py:perf_json.test_compare;run-run_type-upstream-3987: STARTED
 (4/4) perf_json.py:perf_json.test_compare;run-run_type-upstream-3987:  PASS (695.99 s)
RESULTS    : PASS 4 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2024-04-23T14.24-774a6d5/results.html
JOB TIME   : 6972.54 s
```